backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==3.0
pypdfium2==5.3.0
python-dotenv==1.2.1
pyzbar==0.1.9
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
watchdog==6.0.0
